Openosx next

+++ src/Kernel/Extensions/IOPCIFamily/include/IOKit/apiodma/ApplePIODMADefinitions.h
@@ -9,9 +9,9 @@
 #define _IOKIT_ApplePIODMADefinitions_H
 #include <IOKit/IOTypes.h>
 
-#define APIODMABit(bit)                  ((uint32_t)(1) << bit)
-#define APIODMABitRange32(start, end)    (~(((uint32_t)(1) << start) - 1) & (((uint32_t)(1) << end) | (((uint32_t)(1) << end) - 1)))
-#define APIODMABitRange64(start, end)    (~(((uint64_t)(1) << start) - 1) & (((uint64_t)(1) << end) | (((uint64_t)(1) << end) - 1)))
+#define APIODMABit(bit)                  (1u << bit)
+#define APIODMABitRange32(start, end)    ((uint32_t)~((1ull << start) - 1) & ((1ull << end) | ((1ull << end) - 1)))
+#define APIODMABitRange64(start, end)    (~((1ull  << start) - 1) & ((1ull << end) | ((1ull << end) - 1)))
 #define APIODMABitRangePhase(start, end) (start)
 
 #define  kApplePIODMAID "piodma-id"
